sending new queue after game started
