+++ CONFIG.py
@@ -1,2 +1,2 @@
 # If true, make sure to select the LAST role
-TEST_MODE = False
+TEST_MODE = True
