+++ core/match.py
@@ -4,6 +4,7 @@ import json
 import random
 import re
 import traceback
+import uuid
 from datetime import datetime, timedelta
 
 import async_timeout
@@ -21,6 +22,131 @@ LOL_LABELS = ["Top", "Jungle", "Mid", "ADC", "Support"]
 VALORANT_LABELS = ["Controller", "Initiator", "Sentinel", "Duelist", "Flex"]
 OVERWATCH_LABELS = ["Tank", "DPS 1", "DPS 2", "Support 1", "Support 2"]
 OTHER_LABELS = ["Role 1", "Role 2", "Role 3", "Role 4", "Role 5"]
+
+async def start_queue(bot, channel, game, author=None):
+    def region_icon(region, game):
+        if game == "lol":
+            if region == "euw":
+                icon_url = "https://media.discordapp.net/attachments/1046664511324692520/1075444853028175934/OW_Europe.png"
+            elif region == "eune":
+                icon_url = "https://media.discordapp.net/attachments/1046664511324692520/1075444853028175934/OW_Europe.png"
+            elif region == "br":
+                icon_url = "https://media.discordapp.net/attachments/1046664511324692520/1075444852579373136/OW_Americas.png"
+            elif region == "la":
+                icon_url = "https://media.discordapp.net/attachments/1046664511324692520/1075444852579373136/OW_Americas.png"
+            elif region == "jp":
+                icon_url = "https://media.discordapp.net/attachments/1046664511324692520/1075444853233684581/VAL_AP.png"
+            elif region == "las":
+                icon_url = "https://media.discordapp.net/attachments/1046664511324692520/1075444852579373136/OW_Americas.png"
+            elif region == "tr":
+                icon_url = "https://media.discordapp.net/attachments/1046664511324692520/1075444853233684581/VAL_AP.png"
+            elif region == "oce":
+                icon_url = "https://media.discordapp.net/attachments/1046664511324692520/1075444853233684581/VAL_AP.png"
+            elif region == "ru":
+                icon_url = "https://media.discordapp.net/attachments/1046664511324692520/1075444853233684581/VAL_AP.png"
+            else:
+                icon_url = "https://media.discordapp.net/attachments/1046664511324692520/1075444852369670214/VAL_NA.png"
+
+        elif game == "valorant":
+            if region == "ap":
+                icon_url = "https://media.discordapp.net/attachments/1046664511324692520/1077957848161591387/VAL_AP.png"
+            elif region == "br":
+                icon_url = "https://media.discordapp.net/attachments/1046664511324692520/1077957848409067661/VAL_BR.png"
+            elif region == "kr":
+                icon_url = "https://media.discordapp.net/attachments/1046664511324692520/1077957848660713494/VAL_KR.png"
+            elif region == "latam":
+                icon_url = "https://media.discordapp.net/attachments/1046664511324692520/1077957848899801129/VAL_LATAM.png"
+            elif region == "na":
+                icon_url = "https://media.discordapp.net/attachments/1046664511324692520/1077957849130467408/VAL_NA.png"
+            else:
+                icon_url = "https://media.discordapp.net/attachments/1046664511324692520/1075444853028175934/OW_Europe.png"
+
+        elif game == "overwatch":
+            if region == "americas":
+                icon_url = "https://media.discordapp.net/attachments/1046664511324692520/1077957898329673728/OW_Americas.png"
+            elif region == "asia":
+                icon_url = "https://media.discordapp.net/attachments/1046664511324692520/1077957898598101022/OW_Asia.png?width=572&height=572"
+            else:
+                icon_url = "https://media.discordapp.net/attachments/1046664511324692520/1077957898963013814/OW_Europe.png"
+
+        else:
+            icon_url = ""
+        return icon_url
+
+    def banner_icon(game):
+        if game == "lol":
+            return "https://cdn.discordapp.com/attachments/328696263568654337/1068133100451803197/image.png"
+        elif game == "valorant":
+            return "https://media.discordapp.net/attachments/1046664511324692520/1077958380964036689/image.png"
+        elif game == "overwatch":
+            return "https://media.discordapp.net/attachments/1046664511324692520/1077958380636868638/image.png"
+        else:
+            return "https://media.discordapp.net/attachments/328696263568654337/1067908043624423497/image.png?width=1386&height=527"
+
+    def get_title(game):
+        if game == "lol":
+            return "Match Overview - SR Tournament Draft"
+        elif game == "valorant":
+            return "Match Overview - Valorant Competitive"
+        elif game == "overwatch":
+            return "Match Overview - Overwatch Competitive"
+        else:
+            return "Match Overview"
+
+    data = await bot.fetchrow(
+        f"SELECT * FROM queuechannels WHERE channel_id = {channel.id}"
+    )
+    if not data:
+        try:
+            return await channel.send(
+                embed=error(
+                    f"{channel.mention} is not setup as the queue channel, please run this command in a queue channel."
+                )
+            )
+        except:
+            if author:
+                return await author.send(embed=error(f"Could not send queue in {channel.mention}, please check my permissions."))
+
+    # If you change this - update /events.py L28 as well!
+    title = get_title(game)
+    embed = Embed(
+        title=title, color=Color.red()
+    )
+    st_pref = await bot.fetchrow(f"SELECT * FROM switch_team_preference WHERE guild_id = {channel.guild.id}")
+    if not st_pref:
+        embed.add_field(name="Slot 1", value="No members yet")
+        embed.add_field(name="Slot 2", value="No members yet")
+        sbmm = True
+    else:
+        embed.add_field(name="🔵 Blue", value="No members yet")
+        embed.add_field(name="🔴 Red", value="No members yet")
+        sbmm = False
+    if channel.guild.id == 1071099639333404762:
+        embed.set_image(url="https://media.discordapp.net/attachments/1071237723857363015/1073428745253290014/esporty_banner.png")
+    else:
+        banner = banner_icon(game)
+        if banner:
+            embed.set_image(url=banner)
+    with open('assets/tips.txt', 'r') as f:
+        tips = f.readlines()
+        tip = random.choice(tips)
+    embed.set_footer(text="🎮 " +str(uuid.uuid4()).split("-")[0] + '\n' + "💡 " + tip)
+    if not data[1]:
+        data = (data[0], 'na')
+    icon_url = region_icon(data[1], game)
+    if icon_url:
+        embed.set_author(name=f"{data[1].upper()} Queue", icon_url=icon_url)
+    
+    duo_pref = await bot.fetchrow(f"SELECT * FROM duo_queue_preference WHERE guild_id = {channel.guild.id}")
+    if duo_pref:
+        duo = True
+    else:
+        duo = False
+    try:
+        await channel.send(embed=embed, view=Queue(bot, sbmm, duo, game))
+    except:
+        if author:
+            await author.send(embed=error(f"Could not send queue in {channel.mention}, please check my permissions."))
 
 class SpectateButton(ui.View):
     def __init__(self, bot):
@@ -90,7 +216,6 @@ class SpectateButton(ui.View):
     @ui.button(label="Spectate Blue", style=ButtonStyle.blurple, custom_id="lol-specblue")
     async def spec_blue(self, button, inter):
         await self.process_button(button, inter)
-
 
 class RoleButtons(ui.Button):
     def __init__(self, bot, label, custom_id):
@@ -1026,7 +1151,7 @@ class ReadyButton(ui.Button):
                     await self.overwatch_lobby(game_lobby)
 
                 self.disable_button.cancel()
-                # await Match.start(self, inter.channel, self.game)
+                await start_queue(self.bot, inter.channel, self.game)
 
         else:
             await inter.send(
